Update meta.yaml (bypass)
Temporarily changing the source to use the tarball from github instead of from pip 

https://github.com/conda-forge/staged-recipes/blob/main/recipes/example/meta.yaml

so that `requirements_dev.txt` is included. This file wasn't included in the sdist uploaded to pip for v0.3.0, because it wasn't included in the MANIFEST.in file. This has been fixed in the repository, but will be reflected for the next release. 

TODO: change back to https://pypi.io/packages/source/{{ name[0] }}/{{ name }}/{{ name }}-{{ version }}.tar.gz for next release!